+++ sklearn/utils/estimator_checks.py
@@ -1524,7 +1524,8 @@ def check_get_params_invariance(name, estimator):
     if name in ('FeatureUnion', 'Pipeline'):
         e = estimator([('clf', T())])
 
-    elif name in ('GridSearchCV', 'RandomizedSearchCV', 'SelectFromModel'):
+    elif name in ('GridSearchCV', 'RandomizedSearchCV', 'SelectFromModel',
+                  'GPSearchCV'):
         return
 
     else: